added common settings file to accomodate all the settings filea and making translationsettings independent
